post.js file added
